[macOS] Move Waitlist tests into the package as preserved code
Relocates the 6 Waitlist test files from macOS/UnitTests/Waitlist/ into
macOS/LocalPackages/Waitlist/Tests/WaitlistTests/ alongside the code they
test. The Unit Tests target no longer references them (Remove Reference
in Xcode dropped the phantom file refs from project.pbxproj).

The tests only touched package types — @testable import DuckDuckGo_Privacy_Browser
was legacy from when the Waitlist code lived in the main module. Drop it.

The package has no testTarget declared and is not linked by any other
target, so this code is preserved alongside its subject (matching how
iOS treats its Waitlist package) without being compiled in regular
builds or unit test builds.
