+++ macOS/LocalPackages/Waitlist/Tests/WaitlistTests/Mocks/MockWaitlistRequest.swift
@@ -17,7 +17,6 @@
 //
 
 import Foundation
-@testable import DuckDuckGo_Privacy_Browser
 @testable import Waitlist
 
 struct MockWaitlistRequest: WaitlistRequest {
