+++ macOS/LocalPackages/Waitlist/Tests/WaitlistTests/WaitlistViewModelTests.swift
@@ -17,7 +17,6 @@
 //
 
 import XCTest
-@testable import DuckDuckGo_Privacy_Browser
 @testable import Waitlist
 
 final class WaitlistViewModelTests: XCTestCase {
